feat: let env var loading and validation be handled by an app config class
